User ids need to be long

+++ tools/src/main/java/secondbrain/infrastructure/zendesk/api/ZenDeskCommentResponse.java
@@ -3,5 +3,5 @@ package secondbrain.infrastructure.zendesk.api;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record ZenDeskCommentResponse(String body, Integer author_id, String created_at) {
+public record ZenDeskCommentResponse(String body, Long author_id, String created_at) {
 }
